requirements.txt
fastapi
uvicorn[standard]
dnspython
python-multipart
requests
openai>=1.0.0
